XD-1187 Support for --transport for admin
The alias for --transport has been added to xd-admin.
