+++ spring-xd-dirt/src/main/java/org/springframework/xd/dirt/server/options/ContainerOptions.java
@@ -16,7 +16,10 @@
 
 package org.springframework.xd.dirt.server.options;
 
+import javax.validation.constraints.NotNull;
+
 import org.kohsuke.args4j.Option;
+
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
 
@@ -35,6 +38,19 @@ public class ContainerOptions extends CommonDistributedOptions {
 
 	@Option(name = "--hadoopDistro", usage = "The Hadoop distribution to be used for HDFS access")
 	private HadoopDistro distro;
+
+	@Option(name = "--controlTransport", aliases = "--control-transport", usage = "The transport to use for control messages (between admin and nodes)")
+	private ControlTransport controlTransport;
+
+	public void setXD_CONTROL_TRANSPORT(ControlTransport controlTransport) {
+		this.controlTransport = controlTransport;
+	}
+
+	@NotNull
+	public ControlTransport getXD_CONTROL_TRANSPORT() {
+		return controlTransport;
+	}
+
 
 	public void setXD_TRANSPORT(DataTransport transport) {
 		this.transport = transport;
